+++ tests/Fixtures/cache/Nadia-ElasticSearchODM-Tests-Stubs-Document-TestDocument6.dev.php
@@ -9,11 +9,7 @@
   'template' => 
   array (
     'template' => 'idx-testing-*',
-    'settings' => 
-    array (
-      'number_of_shards' => '5',
-    ),
-    'mappings' => 
+    'mappings' =>
     array (
       'log' => 
       array (